worker/pyproject.toml
[tool.pdm]

[project]
name = "worker"
version = "0.0.1"
description = ""
authors = [
    {name = "Nikita Pivkin", email = "delatornikita@gmail.com"},
]
dependencies = [
    "pika>=1.3.1",
    "minio>=7.1.14",
    "python-dotenv>=1.0.0",
    "setuptools>=67.6.1",
    "diffusers>=0.15.0",
    "pip>=23.1",
    "transformers>=4.28.1",
    "accelerate>=0.18.0",
]
requires-python = ">=3.9"
readme = "README.md"

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"
